rename table meta
