+++ vm/src/table_meta.rs
